Course feature: refine openapi
* feat: get course detail


* feat: search course

* feat: get published course

* feat: delete course

* fix: fix name feature

* feat: update basic info

* feat: read model in

* feat: read model golang lint fix

* feat: check backend

* feat(course): gorm slog, and auto migrate

* chore: nothing

* feat: read model in

* feat: read model golang lint fix

* feat: check backend

* feat(course): gorm slog, and auto migrate

* chore(compose): traefik and authentik-blueprints bruh

* chore:  lint for course

* ci: shorten arg

* feat(compose): cors for those services

* fix: cors


* chore: nothing

* fix: update go.mod

* refactor: standardize unbookmark and unhide course terminology

* refactor: update prefix for course

* feat(course): add unbookmark and unhide course endpoints

* refactor(course): change instructorID type from uuid.UUID to string across multiple files

* feat(course): add GetInstructorCourses functionality and update related types

* feat(course): implement enrollment and update course functionalities, add instructor courses retrieval

* feat(course): add finish and review course functionalities, update enrollment handling

* refactor(course): clean up unused code and comments in enrollment and review repositories

* fix(lint): update .redocly.lint-ignore.yaml to include comments for clarity

* chore(compose): authentik api web not need client secret

* fix(compose): authentik more percentage that bootstrap nicely

* refactor(compose): traefik admin inherit instructor, so, no more instructor-or-admin middleware

* feat(course): get video duration from object storage svc

* fix(compose): authentik entitlement identifier with app beside the name

* feat(compose): authentik user app password

* fix(compose): partially missing for previous commit

* chore(repo): constraint slog go lint

* chore(repo): pin golangci-lint version, and no lint for that...

* chore(deps): update dependency @scalar/cli to v1.9.2 (#77)

* chore(api): security scheme add oidc, and scope

* feat: upgrade oapi codegen, nicer to handle error in http controller

* chore(repo): no ffmpeg on windows

* feat(repo): mockery, test, and unauthorize in api just object, no detail inside

* feat(repo): agents skill for testing

* chore(repo): test-go skill adjust abit style

* feat(course): seed

* feat(course): doing

* feat(course): we are going tooooooo

* test

* feat(repo): script to download seed videos

* fix(course): grpc bruh

* chore:

* chore: test pre commit

* refactor(api/course): path prefix those aggreate, but not params yet

* feat(api/course): params aggreate

* fix(repo): mise ffmpeg no lock version, and js runtime node for ytdlp

* fix(compose): traefik another way to protect

* chore(api): remove obsolete un[AZ]

* chore(deps): update dependency @hey-api/openapi-ts to v0.97.1

* chore(deps): update traefik docker tag to v3.6.16

* fix(deps): update dependency @scalar/api-reference-react to v0.9.34

* chore(deps): update dependency vitepress-sidebar to v1.33.2

* refactor: remove unbookmark and unhide course functionalities

* feat: rename and implement getMyCourses endpoint

* Refactor course API: Remove searchCourses endpoint and related types; update course models to use introductionVideoUrl; add instructorId query parameter for published courses.

* feat: add instructorID to GetCourses and set IntroductionVideoKey to nil in DTOs

* Refactor (course API): Remove test lesson progress and upload video lesson URL endpoints; add new upload video URL endpoint with updated request and response structures.

* Refactor: Remove LessonID from GetUploadVideoLessonURLParams and related error handling; update NewVideoKeyParams accordingly.

* feat: Add course review and analytics features

* helppp

* chore: ok what? missing wire gen rename

---------

+++ internal/course/app/cmd_updatecourse.go
@@ -55,7 +55,7 @@ func (h *UpdateCourseHandler) Handle(ctx context.Context, cmd *UpdateCourse) err
 			}
 		}
 		if cmd.Price != 0 {
-			if err := course.SetPrice(float64(cmd.Price)); err != nil {
+			if err := course.SetPrice(cmd.Price); err != nil {
 				return err
 			}
 		}
